backend/requirements.txt
Flask==3.1.*
Flask-SQLAlchemy==3.1.*
Flask-CORS==5.*
PyJWT==2.*
Werkzeug==3.*
python-dotenv==1.*
